chore: move chain state notifications to reth-chain-state crate
