+++ crates/storage/provider/Cargo.toml
@@ -39,7 +39,6 @@ revm.workspace = true
 
 # async
 tokio = { workspace = true, features = ["sync", "macros", "rt-multi-thread"] }
-tokio-stream = { workspace = true, features = ["sync"] }
 
 # tracing
 tracing.workspace = true
@@ -50,9 +49,7 @@ metrics.workspace = true
 
 # misc
 auto_impl.workspace = true
-derive_more.workspace = true
 itertools.workspace = true
-pin-project.workspace = true
 parking_lot.workspace = true
 dashmap = { workspace = true, features = ["inline"] }
 strum.workspace = true
@@ -78,4 +75,9 @@ rand.workspace = true
 [features]
 optimism = ["reth-primitives/optimism", "reth-execution-types/optimism"]
 serde = ["reth-execution-types/serde"]
-test-utils = ["alloy-rlp", "reth-db/test-utils", "reth-nippy-jar/test-utils"]
+test-utils = [
+  "alloy-rlp",
+  "reth-db/test-utils",
+  "reth-nippy-jar/test-utils",
+  "reth-chain-state/test-utils"
+]
