+++ crates/storage/provider/src/test_utils/mod.rs
@@ -7,13 +7,12 @@ use reth_db::{
 use std::sync::Arc;
 
 pub mod blocks;
-mod events;
 mod mock;
 mod noop;
 
-pub use events::TestCanonStateSubscriptions;
 pub use mock::{ExtendedAccount, MockEthProvider};
 pub use noop::NoopProvider;
+pub use reth_chain_state::test_utils::TestCanonStateSubscriptions;
 
 /// Creates test provider factory with mainnet chain spec.
 pub fn create_test_provider_factory() -> ProviderFactory<Arc<TempDatabase<DatabaseEnv>>> {
