+++ crates/storage/provider/src/traits/tree_viewer.rs
@@ -1,5 +1,6 @@
-use crate::{BlockchainTreePendingStateProvider, CanonStateSubscriptions};
+use crate::BlockchainTreePendingStateProvider;
 use reth_blockchain_tree_api::{BlockchainTreeEngine, BlockchainTreeViewer};
+use reth_chain_state::CanonStateSubscriptions;
 
 /// Helper trait to combine all the traits we need for the `BlockchainProvider`
 ///
